+++ CHANGELOG.md
@@ -25,8 +25,10 @@ handleCliError(error, { stack: false })
   colors. It defaults to `true` in terminals.
 - Quoted strings in the error message are now
   [colorized](README.md#%EF%B8%8F-colors).
-- An icon is now prepended to the error's name. It can be configured using the
+- An icon is now prepended to the error's name. This can be configured using the
   [`icon` option](README.md#-icon)
+- The error's icon and name are now logged in red by default. This can be
+  configured using the [`header` option](README.md#-header)
 - The [`props` option](README.md#-props) has been added to hide error properties
 
 # 2.5.1